get build to work after ensuring incoming and outgoing control packets are getting sent

+++ client/src/commonMain/kotlin/mqtt/client/session/SessionState.kt
@@ -17,7 +17,7 @@ class ClientSessionState(
     val qos2MessagesRecevedButNotCompletelyAcked: QueuedObjectCollection = MemoryQueuedObjectCollection()
 ) {
     val subscriptionManager = SubscriptionManager()
-    val unacknowledgedSubscriptions = HashMap<UShort, ISubscribeRequest>()
+    val unacknowledgedSubscriptions = HashMap<Int, ISubscribeRequest>()
 
     suspend fun start(clientId: MqttUtf8String, server: Url) {
         qos1And2MessagesSentButNotAcked.open(clientId, server)
