+++ wire-core/src/commonMain/kotlin/mqtt/wire/control/packet/PublishComplete.kt
@@ -3,5 +3,5 @@
 package mqtt.wire.control.packet
 
 interface IPublishComplete : ControlPacket {
-    val packetIdentifier: UShort
+    val packetIdentifier: Int
 }
