+++ wire-core/src/commonMain/kotlin/mqtt/wire/control/packet/SubscribeRequest.kt
@@ -5,7 +5,7 @@ package mqtt.wire.control.packet
 import mqtt.wire.data.topic.Filter
 
 interface ISubscribeRequest : ControlPacket {
-    val packetIdentifier: UShort
+    val packetIdentifier: Int
     fun expectedResponse(): ISubscribeAcknowledgement
     fun getTopics(): List<Filter>
 }
