+++ wire4/src/commonTest/kotlin/mqtt/wire4/control/packet/PublishAcknowledgement.kt
@@ -6,7 +6,7 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class PublishAcknowledgementTest {
-    private val packetIdentifier = 2.toUShort()
+    private val packetIdentifier = 2
 
     @Test
     fun packetIdentifier() {
