+++ wire4/src/commonTest/kotlin/mqtt/wire4/control/packet/UnsubscribeRequest.kt
@@ -8,7 +8,7 @@ import kotlin.test.assertEquals
 
 class UnsubscribeRequestTests {
 
-    private val packetIdentifier = 2.toUShort()
+    private val packetIdentifier = 2
 
     @Test
     fun basicTest() {
